old/requirements.txt
discord.py==1.7.1
asteval==0.9.23
ISStreamer==1.0.1
numpy==1.20.2
azure
azure-keyvault-secrets==4.2.0
azure-identity==1.5.0
pymongo==3.11.3
motor==2.4.0
psutil==5.8.0
Whoosh==2.7.4
aioodbc==0.3.3
